+++ oneflow/init.py
@@ -25,7 +25,7 @@ del session_util
 
 import oneflow.python.framework.register_python_callback
 
-import oneflow.python.__export_symbols__
+import oneflow.python_gen.__export_symbols__
 
 import atexit
 import oneflow.python.framework.c_api_util
@@ -37,3 +37,4 @@ del atexit
 del absolute_import
 del oneflow
 del python
+del python_gen
